Tests/LibWeb/Text/expected/content-blocker-svg-image-skips-cosmetic-css.txt

html display: none
svg pixel: 255,0,0,255
